tests/wpt/web-platform-tests/webxr/webGLCanvasContext_create_xrcompatible.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>
  <script src="resources/webxr_test_constants.js"></script>
  <script>

    function testNoDevice(t, gl) {
      assert_false(gl.getContextAttributes().xrCompatible);
      return promise_rejects_dom(t, "InvalidStateError", gl.makeXRCompatible());
    }

    xr_promise_test("Creating a webgl context with no device",
      testNoDevice, null, 'webgl');

    xr_promise_test("Creating a webgl2 context with no device",
      testNoDevice, null, 'webgl2');

    function testXrCompatible(t, gl) {
      let offscreenGl = null;
      let offscreenGl2 = null;
      return navigator.xr.test.simulateDeviceConnection(TRACKED_IMMERSIVE_DEVICE)
        .then( (controller) => {
          return gl.makeXRCompatible();
        }).then( () => {
          assert_true(gl.getContextAttributes().xrCompatible);

          // Check that an offscreen webgl context behaves no different.
          let offscreenCanvas = document.createElement('canvas');
          offscreenGl = offscreenCanvas.getContext('webgl');
          return offscreenGl.makeXRCompatible();
        }).then( () => {
          assert_true(offscreenGl.getContextAttributes().xrCompatible);

          // Check that an offscreen webgl2 context behaves no different.
          let offscreenCanvas = document.createElement('canvas');
          offscreenGl2 = offscreenCanvas.getContext('webgl2');
          return offscreenGl2.makeXRCompatible();
      }).then( () => {
          assert_true(offscreenGl2.getContextAttributes().xrCompatible);
      });
    }

    xr_promise_test("An XR-compatible webgl context can be created",
      testXrCompatible, null, 'webgl');

    xr_promise_test("An XR-compatible webgl2 context can be created",
      testXrCompatible, null, 'webgl2');

  </script>
</body>
